mcp_streamlit/requirements.txt
streamlit==1.36.0
pydantic==2.7.4
pydantic-settings==2.3.4
PyYAML==6.0.1
pyyaml>=6,<7
httpx==0.27.0
numpy==1.26.4
pandas==2.2.2
networkx==3.3
pyvis==0.3.2
rank-bm25==0.2.2
tqdm==4.66.4
rapidfuzz>=3.9,<4
